Synthetic_backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
mysql-connector-python==8.1.0
pandas==2.2.3
matplotlib==3.8.2
sdv==1.1.0
sdv[synthesizers]==1.1.0
pyjwt==2.8.0
Werkzeug==3.0.0
